+++ apps/studio/components/interfaces/Auth/Users/UserOverview.tsx
@@ -21,7 +21,7 @@ import { timeout } from 'lib/helpers'
 import { Button, cn, Separator } from 'ui'
 import { Admonition } from 'ui-patterns/admonition'
 import ConfirmationModal from 'ui-patterns/Dialogs/ConfirmationModal'
-import { PROVIDERS_SCHEMAS } from '../AuthProvidersFormValidation'
+import { PROVIDERS_SCHEMAS } from '../AuthProvidersForm/AuthProvidersFormValidation'
 import { BanUserModal } from './BanUserModal'
 import { DeleteUserModal } from './DeleteUserModal'
 import { UserHeader } from './UserHeader'
